feat(zbrowser): rewrite auth, implement CRUD, fix bugs
- Refactor auth: split Cookie into guid/mid/Q/T fields
- Assemble Cookie header and query params in apiRequest
- Implement Rename (/v3/file/rename)
- Implement Copy (/v3/dir/copy)
- Implement Remove (/v3/recycle/move, /v3/recycle/del) with DeleteMode option
- Fix Link() referer calculation (strings.LastIndex)
- Fix Link() Files[0] bounds check
- Fix Link() pid to use actual parent ID
- Fix List() pagination infinite loop guard
- Fix Copy() return value (nil instead of source object)
- Fix dirNewRespV3 timestamp types (int -> int64)
- Add client nil defensive check in apiRequest
- Improve HTTP error messages with response body
- Remove duplicate Content-Type header from Init
- Set all ver values to 0

+++ drivers/zbrowser/types.go
@@ -138,12 +138,12 @@ type dirNewRespV3 struct {
 			DirName      string `json:"dirName"`
 			Pid          string `json:"pid"`
 			Status       int    `json:"status"`
-			CollectAt    int    `json:"collectAt"`
-			DelAt        int    `json:"delAt"`
+			CollectAt    int64  `json:"collectAt"`
+			DelAt        int64  `json:"delAt"`
 			CollectDirID int    `json:"collectDirId"`
 			Ver          int    `json:"ver"`
-			CreatedAt    int    `json:"createdAt"`
-			UpdatedAt    int    `json:"updatedAt"`
+			CreatedAt    int64  `json:"createdAt"`
+			UpdatedAt    int64  `json:"updatedAt"`
 		} `json:"model"`
 	} `json:"data"`
 }
